Parsing File and Recurring Event Creator File
FileParser - parses a given text file and returns ArrayList<String[]>. Simply parses, and does not check if inputs are valid or if there are enough inputs.
RecurringEventCreator - creates ArrayList<Event> of recurring events. Checks if enough parameters are given for each String[] in the Arraylist. Does not check for overlapping events. Code can be edited incorporated into another file that handles the creation of events (and the checking of overlapping events).
NOTE: RecurringEventCreator has not been fully tested.